backend/requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
pkg-resources==0.0.0
promise==2.3
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2020.5
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
